requirements.txt
Flask==3.0.3
hugchat==0.4.6
openai==1.30.4
python-dotenv==1.0.1
